Modificacoes no modelo. Metodo para executar votacao. Metodo para cadastrar uma pauta.

+++ src/main/java/south/system/test/sessaovotacao/repository/IPautaRepository.java
@@ -3,5 +3,5 @@ package south.system.test.sessaovotacao.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import south.system.test.sessaovotacao.model.Pauta;
 
-public interface IPautaRepository extends JpaRepository<Pauta, Integer> {
+public interface IPautaRepository extends JpaRepository<Pauta, Long> {
 }
